api/requirements.txt
azure-functions==3.2.0
pandas==1.3.5
numpy==1.21.4
yfinance==0.1.70
python-dateutil==2.8.2